Move server plugins
