+++ packages/react-cosmos/src/server/corePlugins/index.ts
@@ -1,4 +1,4 @@
-import { CosmosServerPlugin } from '../../cosmosPlugin/types.js';
+import { CosmosServerPlugin } from '../cosmosPlugin/types.js';
 import { httpProxyServerPlugin } from './httpProxy.js';
 import { openFileServerPlugin } from './openFile.js';
 import { pluginEndpointServerPlugin } from './pluginEndpoint.js';
